rename file GenAi
